fix: writes after namespace created
Before this, the `create_wal_directory` would run
and panic because no notion of the directoy already
being created is tracked.

This changes the `replay` function such that it is
no longer independent of the `NamespaceWal` and
instead populates the internal `namespaces` mapping
of known namespaces as the replay occurs.