refactor: use the log (#41)
* chore: remove/stub in-memory queue paths

This removes the in-memory `Vec` for messages
that have been received and stubs out various paths
with `todo!` and `unimplemented!` in place so that
the code still compiles.

However, tests fail which is the main aim. Once these
tests pass again, this should be ready for merging
into main.

* refactor: segment tracks it own id

* refactor: remove namespace appends from wal

A namespace existing in a directory is evidence of its creation, this is already on disk

* feat: add namespace config type

* refactor: remove id from individual wal

This means that the relevant segment id must be found during the creation of a wal

* fix: dont set segment twice

* feat: find segment ids instead of static construction

* refactor: return offset/write size

* docs: expand design

* refactor: wal wip

* refactor: write/read from log

Write a non varint encoded entry and use a regular u64 length delimiter that we can read.

This also includes accompanying tests for reading

* feat: include logical offset to encoded message

* fix: tests for wal

* refactor: add read/get_messages impls

* fix: add flush per write

This ensures writes land on disk, beforehand they were served out of the in-memory representation even if they hadn't made it to disk yet

* feat: create/contains namespace

* chore: snip unused test for updating, comment replay

* chore: clippy

Address clippy lints which are not for replays,
this is going to be handled separately.

* refactor: rename to scoped/multi logs

* chore: clippy

* docs: scoped log

* chore: explicit segment id usage

* docs: log term

* feat: add scoped log replays

* chore: make directory no-op if exists

* feat: use histogram w/ timer for wal_replay_duration

* feat: add multilog replay fn

* chore: clippy